requirements.txt
streamlit
boto3
pandas
Pillow